requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.0
streamlit==1.32.0
requests==2.31.0
pydantic==2.5.3
